cloud_core/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
redis>=5.0.7
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
numpy>=1.26.0
scikit-learn>=1.5.0
prometheus-client>=0.22.0
